k3y_2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
email-validator
sendgrid
pydantic
python-dotenv
aiofiles
python-dotenv
